cbt_chatbot/pyproject.toml
[tool.poetry]
name = "cbt-chatbot"
version = "0.1.0"
description = "This is our integrated therapy software code"
authors = ["GrantGuernsey <78874644+GrantGuernsey@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
Django = "^5.1.2"
openai = "^1.53.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
